Record Methode operation metrics per host based on IP address.
Add a method to MethodeObjectFactory to provide the IP address of the
Methode host. In the default implementation the IOR is then set to use
the IP address, so the call will definitely be directed to that host for
the duration of the call even if DNS switches over immediately
afterwards.

Enhance MethodeRepositoryOperationTemplate (and other templates that
delegate to it likewise) by allowing it to accept a class name and timer
name in the constructor, and using these values in combination with the
Methode IP address to define a Dropwizard timer against which calls can
be recorded. Use the new constructor in operations in
MethodeFileRepository, which will then emit metrics such as
"com.ft.methodeapi.service.methode.MethodeFileRepository":
{"findFileByUuid@10.112.36.81": {...} }. If the DNS resolution swaps to
a different IP address (i.e. data centre failover), then the metrics for
these calls will be recorded under a different key based on the IP
address, so any performance divergence can be identified.